requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
idna==3.11
Markdown==3.10.2
pillow==12.2.0
PyYAML==6.0.3
requests==2.33.1
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
